+++ src/Ironforge/UserBundle/Repository/UserRepository.php
@@ -24,4 +24,27 @@ class UserRepository extends EntityRepository
 
         return $query->getSingleScalarResult();
     }
+
+    /**
+     * @param string $order
+     * @param int    $limit
+     *
+     * @return int
+     */
+    public function getSortedUserByUnlockedBadges($order = 'DESC', $limit = 7)
+    {
+        $qb = $this->getEntityManager()->createQueryBuilder();
+        $qb->select('u AS user, COUNT(ub.id) AS nbUnlockedBadges')
+            ->from('UserBundle:User', 'u')
+            ->leftJoin('AchievementBundle:UnlockedBadge', 'ub')
+            ->where('ub.user = u')
+            ->setMaxResults($limit)
+            ->orderBy('nbUnlockedBadges', $order)
+            ->groupBy('u')
+        ;
+
+        $query = $qb->getQuery();
+
+        return $query->getResult();
+    }
 }